BananaDoc_AI/requirements.txt
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
Flask==2.3.2
matplotlib==3.7.1 